kamp-core/src/main/kotlin/ch/leadrian/samp/kamp/core/runtime/callback/CheckpointCallbackListener.kt
package ch.leadrian.samp.kamp.core.runtime.callback

import ch.leadrian.samp.kamp.core.api.callback.CallbackListenerManager
import ch.leadrian.samp.kamp.core.api.callback.OnPlayerEnterCheckpointListener
import ch.leadrian.samp.kamp.core.api.callback.OnPlayerLeaveCheckpointListener
import ch.leadrian.samp.kamp.core.api.entity.Player
import javax.annotation.PostConstruct
import javax.inject.Inject
import javax.inject.Singleton

@Singleton
internal class CheckpointCallbackListener
@Inject
constructor(
        private val callbackListenerManager: CallbackListenerManager
) : OnPlayerEnterCheckpointListener, OnPlayerLeaveCheckpointListener {

    @PostConstruct
    fun initialize() {
        callbackListenerManager.register(this)
    }

    override fun onPlayerEnterCheckpoint(player: Player) {
        player.checkpoint?.onEnter(player)
    }

    override fun onPlayerLeaveCheckpoint(player: Player) {
        player.checkpoint?.onLeave(player)
    }

}